cpberc agregado de confguracion y preparacion de jsp plantillas

+++ src/main/java/dao/IGenericDAO.java
@@ -8,13 +8,13 @@ public interface IGenericDAO<T> {
 
 	List<T> getAll() throws PersistenceException;
 	
-	T getByID(long id) throws PersistenceException;
+	T getByID(int id) throws PersistenceException;
 	
 	T save(T entidad) throws PersistenceException;
 	
 	T update(T entidad) throws PersistenceException;
 	
-	boolean delete( long id ) throws PersistenceException;
+	boolean delete( int id ) throws PersistenceException;
 	
 	boolean delete( T entidad ) throws PersistenceException;
